requirement.txt
azureml-sdk
mlflow
pandas
numpy
scikit-learn
matplotlib
seaborn
plotly